fixing mergeVar param of Subscribe method
Changed the type of the mergeVar param of the Subscribe method to be an
object instead of MergeVar so that any serializable object can be passed
in.  This allows the user to derive from the MergeVars class and define
any custom merge variables they want to post to the Subscribe api call.
Also set JsConfig.ExcludeTypeInfo to true to supress the "__type" member
of the emitted JSON from appearing.